requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.3
dnspython==2.2.1
fastapi==0.85.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
jwt==1.3.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.8.0
pymongo==4.3.2
python-dotenv==0.21.0
python-slugify==6.1.2
sniffio==1.3.0
starlette==0.20.4
text-unidecode==1.3
typing_extensions==4.4.0
uvicorn==0.19.0
